+++ examples/vnext/mastra.docker-compose.yaml
@@ -1,9 +1,9 @@
 services:
   db:
-    image: postgres:13
+    image: pgvector/pgvector:pg16
     container_name: 'vnext-db'
     ports:
-      - '5433:5433'
+      - '5433:5432'
     environment:
       POSTGRES_USER: ${POSTGRES_USER:-postgres}
       POSTGRES_PASSWORD: ${POSTGRES_PASSWORD:-postgres}
