+++ packages/core/src/sync/types.ts
@@ -1,6 +1,7 @@
 import { Agent } from "../agent";
 import { MastraEngine } from "../engine";
 import { Integration } from "../integration";
+import { LLM } from "../llm";
 import { AllTools, ToolApi } from "../tools/types";
 import { MastraVector } from "../vector";
 
@@ -17,6 +18,7 @@ export type SyncConfig<
     engine: MastraEngine;
     vectors?: Record<string, MastraVector>;
     agents: Map<string, Agent<TIntegrations, MastraTools>>;
+    llm: LLM<TIntegrations, MastraTools, keyof AllTools<TIntegrations, MastraTools>>;
 }) => Promise<void>;
 
 export interface SyncMap {
